+++ CHANGELOG.md
@@ -12,7 +12,8 @@ All notable changes to this project will be documented in this file, in reverse 
 - PSR-7 double pass is now supported via [tuupola/callable-handler](https://github.com/tuupola/callable-handler) library.
 - PHP 7.1 is now minimal requirement.
 - Error callback now receives only response and arguments, request was removed.
-- Before callback now receives only request and arguments
+- Before callback now receives only request and arguments, response was removed.
+- After callback now receives only response and arguments, request was removed.
 
 ### Added
 - Support for the [approved version of PSR-15](https://github.com/php-fig/http-server-middleware).